Include User Info in log messages.
Formerly most of the log messages were using `localizedDescription`, but that is intended for display to the user in the UI in their locale, not for use in engineering logs.

This is particularly necessary for Codable errors, which include details as to why the failure occurred.

This was already done in PersistedValue, but should have been done everywhere.