Include User Info in log messages. (#94)
Formerly most of the log messages were using `localizedDescription`, but that is intended for display to the user in the UI in their locale, not for use in engineering logs.

This is particularly necessary for Codable errors, which include details as to why the failure occurred.  I've wanted to make this change in Campaigns for a while, as hidden errors invariably trip me up every time.  This change has already been done to the main Rover SDK.

It was already done in PersistedValue, but should have been done everywhere.